Moved components folder to root path
